Migrate Cilium to install/uninstall via new CLI
The Cilium community has developed a new CLI for installing, operating
and debugging Kubernetes clusters with Cilium installed[0]. We can reuse
that CLI here for a much simpler, less error-prone upgrade process.

For the most part, the new CLI almost removes the need for these
microk8s-specific scripts. However, the CLI cannot know how to uninstall
existing CNIs in a microk8s-specific environment, so we still need some
logic here to clean that up in order for Cilium to successfully deploy
and avoid conflicts with existing components.

A few notes about implementation choices:
* Remove previous CNIs and CNI config before initiating Cilium install
  to ensure that Cilium manages pods after the cilium install command
  has been run.
* disable.cilium.sh is handling disabling the old version during
  microk8s upgrade.
* Use 'rm -rf -- $directories" to reduce likelihood of removing bad
  directories, rather than 'rm -rf $directories'

[0] https://github.com/cilium/cilium-cli/

+++ microk8s-resources/actions/disable.cilium.sh
@@ -6,10 +6,15 @@ source $SNAP/actions/common/utils.sh
 
 echo "Disabling Cilium"
 
-"$SNAP/kubectl" "--kubeconfig=$SNAP_DATA/credentials/client.config" delete -f "$SNAP_DATA/actions/cilium.yaml"
+# Upgrade from v1.21 or earlier.
+if [ -e "$SNAP_DATA/actions/cilium.yaml" ]; then
+  "$SNAP/kubectl" "--kubeconfig=$SNAP_DATA/credentials/client.config" delete -f "$SNAP_DATA/actions/cilium.yaml"
+  # Give K8s some time to process the deletion request
+  sleep 15
+else
+  KUBECONFIG="$SNAP_DATA/credentials/client.config" "$SNAP_DATA/bin/cilium" uninstall --wait
+fi
 
-# Give K8s some time to process the deletion request
-sleep 15
 cilium=$(wait_for_service_shutdown "kube-system" "k8s-app=cilium")
 if [[ $cilium == fail ]]
 then
@@ -21,14 +26,19 @@ if [[ $cilium == fail ]]
 then
   echo "Cilium operator did not shut down on time. Proceeding."
 fi
+
+# These commands handle upgrade from Microk8s v1.21 or earlier.
 run_with_sudo rm -f "$SNAP_DATA/args/cni-network/05-cilium-cni.conf"
 run_with_sudo rm -f "$SNAP_DATA/opt/cni/bin/cilium-cni"
-run_with_sudo rm -rf $SNAP_DATA/bin/cilium*
+run_with_sudo rm -rf -- $SNAP_DATA/bin/cilium-*
 run_with_sudo rm -f "$SNAP_DATA/actions/cilium.yaml"
-run_with_sudo rm -rf "$SNAP_DATA/actions/cilium"
-run_with_sudo rm -rf "$SNAP_DATA/var/run/cilium"
-run_with_sudo rm -rf "$SNAP_DATA/sys/fs/bpf"
+run_with_sudo rm -rf -- "$SNAP_DATA/actions/cilium"
+run_with_sudo rm -rf -- "$SNAP_DATA/sys/fs/bpf"
+# End Microk8s v1.21 or earlier block
 
+run_with_sudo rm -f "$SNAP_DATA/args/cni-network/05-cilium.conf"
+run_with_sudo rm -f "$SNAP_DATA/bin/cilium"
+run_with_sudo rm -rf -- "$SNAP_DATA/var/run/cilium"
 if $SNAP/sbin/ip link show "cilium_vxlan"
 then
   echo "Deleting old cilium_vxlan link"
